backend/requirements.txt
asgiref==3.6.0
autopep8==2.0.1
contourpy==1.0.7
cycler==0.11.0
Django==4.1.6
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
fonttools==4.38.0
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.3
numpy==1.24.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pycodestyle==2.10.0
pymongo==3.12.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
scikit-learn==1.2.1
scipy==1.10.0
six==1.16.0
sqlparse==0.2.4
threadpoolctl==3.1.0
tomli==2.0.1
